% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{calc_AUC_lag}
\alias{calc_AUC_lag}
\title{calculate AUC values from Lag phase data}
\usage{
calc_AUC_lag(lag_data)
}
\arguments{
\item{lag_data}{Output of calc_lag_phase() function}
}
\value{
Outputs AUC lag-phase-curve values for each sample with matching sample info
}
\description{
Calculates lag phase AUC for each sample from the lag phase data. Converts dilution to -log10(dilution) for AUC calculation
}
\examples{
auc_res <- calc_AUC(lag_data)
}
