% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{calc_AUC_sig}
\alias{calc_AUC_sig}
\title{calculate AUC values from signal curves}
\usage{
calc_AUC_sig(sig_data)
}
\arguments{
\item{sig_data}{Output of signal_curve() function}
}
\value{
Outputs AUC values from signal curves for each well with matching sample info
}
\description{
Calculates lag phase AUC for each well from the signal curve.
}
\examples{
auc_res <- calc_AUC(sig_data)
}
