% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{calc_lag_phase}
\alias{calc_lag_phase}
\title{calculate 1/lag phase}
\usage{
calc_lag_phase(
  data,
  sample_info,
  cutoff,
  thresh_method = "StdDev",
  threshold,
  thresh_calc_range = c(1:4)
)
}
\arguments{
\item{data}{formatted RT-QuIC fluorescence data. Output of 'load_quic_results()'}

\item{sample_info}{Matrix that contains matching sample information for each well in the matching QuIC plate}

\item{cutoff}{The time cutoff in hours that will serve as the max time to reach threshold fluorescence}

\item{thresh_method}{The method for calculating threshold fluorescence. Default is "StdDev" (based on standard deviation). May also use "2xMean" (2xmean(fluourescence)), "Max" (10\% of max fluorescence) or "Manual" (manually specify a number).}

\item{threshold}{Numerical value to be used as threshold. Only applies when thresh_method = "Manual"}

\item{thresh_calc_range}{Range of cycle numbers to be used as the basis for calculating threshold. By default thresh_calc_range = c(1:4).}
}
\value{
Outputs a matrix with the 1/lag values for each well with matching sample info data
}
\description{
calculates the inverse lag phase for each well of the QuIC plate (AKA amyloid formation rate)
}
\examples{
lag_res <- calc_lag_phase(res, samples, 40)
lag_res <- calc_lag_phase(res, samples, 40, thresh_method = "2xMean", thresh_calc_range = c(13:16))
lag_res <- calc_lag_phase(res, samples, 40, thresh_method = "Manual", threshold=15000)
}
