% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{calc_mpr}
\alias{calc_mpr}
\title{calculate max-point fluorescence ratios (MPR)}
\usage{
calc_mpr(data)
}
\arguments{
\item{data}{formatted RT-QuIC fluorescence data. Output of 'signal_curve'
#' @examples
mpr_res <- calc_mpr(signal_data)}
}
\value{
Outputs a matrix with the MPR values for each well with matching sample info data
}
\description{
calculate max-point fluorescence ratios (MPR)
}
