% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{calc_SD50}
\alias{calc_SD50}
\title{calculate SD50 values}
\usage{
calc_SD50(lag_data, starting_dilution = NULL, positivity_threshold = 0.75)
}
\arguments{
\item{lag_data}{Output of \code{calc_lag_phase()} with columns: Well, lag_time, Sample, Dilution}

\item{starting_dilution}{Optional numeric. Use only dilutions ≤ this value (e.g., for inhibition at high concentration)}

\item{positivity_threshold}{Proportion of positive wells required in highest dilution. Default = 0.75}
}
\value{
Data frame with SD50 metrics and result interpretation
}
\description{
Calculates SD50 for each sample from the lag phase data. 3/4 wells for the highest dilution must be
positive for the function to return SD50 measurements (This serves as the limit of detection).
Otherwise returns the maximum possible value below the limit of detection.
}
\examples{
\dontrun{sd50_res <- calc_SD50(lag_data)}
\dontrun{sd50_res <- calc_SD50(lag_data, starting_dilution=1e-04)}
}
