% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{convert_time_values}
\alias{convert_time_values}
\title{Convert time values}
\usage{
convert_time_values(QuIC_data)
}
\arguments{
\item{QuIC_data}{formatted QuIC results matrix with funky time values}
}
\value{
Matrix with proper numerical values for time in hours
}
\description{
Converts time values from a funky string of text to proper numerical values in hours.
Used by the load_quic_results() function
}
