% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{load_quic_results}
\alias{load_quic_results}
\title{Load and format data frame}
\usage{
load_quic_results(
  input_file,
  file_type = "clean_table",
  excel_sheet = NULL,
  rows_to_skip = NULL,
  unnecessary_columns = NULL
)
}
\arguments{
\item{input_file}{the path to the file containing raw RT-QuIC fluorescence data values in microsoft excel format (.xslx)}

\item{file_type}{How is the input file formatted? By default, file_type = "clean_table". file_type = "raw_table" for raw data off of the omega in table format.  file_type = "raw_microplate" for raw data off of the omega in microplate format.}

\item{excel_sheet}{(if more than one sheet) The name of the excel sheet that contains the formatted data}

\item{rows_to_skip}{(file_type = "raw_table" and file_type = "raw_microplate") The number of header rows to skip in the excel sheet before the data table begins.}

\item{unnecessary_columns}{(file_type = "raw_table" ONLY) The column number(s) of any unnecessary columns in the excel sheet. The only columns should be the Well ID, and individual reads.}
}
\value{
Properly formatted matrix of QuIC data
}
\description{
Takes raw data and formats it for further analysis
}
\examples{
\dontrun{load_quic_results("raw data/quic_results.xlsx")}
\dontrun{load_quic_results("raw data/quic_results.xlsx", filetype="raw_table", excel_sheet="All Cycles", rows_to_skip = 12, unnecessary_columns = c(2,3))}
}
