% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{plot_lag_phase}
\alias{plot_lag_phase}
\title{generic plot of lag phase data}
\usage{
plot_lag_phase(lag_data)
}
\arguments{
\item{lag_data}{Output of calc_lag_phase() function}
}
\description{
Makes a generic plot of lag phase data for each QuIC plate
}
\examples{
\dontrun{plot_lag_phase(lag_data)}
}
