% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{plot_signal_curve}
\alias{plot_signal_curve}
\title{generic signal curve plot}
\usage{
plot_signal_curve(plot_data)
}
\arguments{
\item{plot_data}{output of 'signal_curve()' function}
}
\description{
makes a generic plot with ggplot from the formatted signal curve plot data
}
\examples{
plot_signal_curve(plot_data)
}
