% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{signal_curve}
\alias{signal_curve}
\title{Format data for signal curve plot}
\usage{
signal_curve(
  plot_data,
  plot_samples,
  normalize = "none",
  baseline_cycles = NULL,
  smooth = FALSE
)
}
\arguments{
\item{plot_data}{The formatted QuIC fluorescence data that will be used for plotting. Output of 'load_quic_results()'}

\item{plot_samples}{Matrix that contains matching sample information for each well in the matching QuIC plate}

\item{normalize}{Normalize method for fluorescence signal. Either normalize="none", normalize="max_RFU_per_plate", or normalize="baseline_RFU_per_well"}

\item{baseline_cycles}{Must be set only when normalize="baseline_RFU_per_well". Cycles used for baseline calculation. e.g. baseline_cycles=c(13:16)}

\item{smooth}{Whether to smooth the fluorescence curves. Needed when calculating lag times from slopes.}
}
\value{
Formatted matrix of fluorescence for each well with sample info for signal curve plotting
}
\description{
Takes matrix of QuIC fluorescence values and formats it with sample info to plot signal against time with ggplot
}
\examples{
\dontrun{sig_res <- signal_curve(QuIC_res, samples, normalize = "max_RFU_per_plate")}
\dontrun{sig_res <- signal_curve(QuIC_res, samples, normalize = "baseline_RFU_per_well", baseline_cycles = c(13:16), smooth = TRUE)}
}
