% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{calc_lag_phase}
\alias{calc_lag_phase}
\title{calculate 1/lag phase}
\usage{
calc_lag_phase(
  signal_data,
  cutoff,
  thresh_method = "StdDev",
  n_StdDevs = 10,
  mean_FC = 2,
  proportion_max = 0.1,
  threshold,
  thresh_calc_range = c(0, 1)
)
}
\arguments{
\item{signal_data}{output of 'signal_curve()' function}

\item{cutoff}{The time cutoff in hours that will serve as the max time to reach threshold fluorescence}

\item{thresh_method}{The method for calculating threshold fluorescence. Default is "StdDev" (based on standard deviation). May also use "Mean" (2xmean(fluourescence)), "Max" (10\% of max fluorescence), "Manual" (manually specify a number) or "MaxSlope" (based on curve slope instead of signal threshold).}

\item{n_StdDevs}{Only applies when thresh_method = "StdDev"; The number of standard deviations above baseline for thresholding (10 by default).}

\item{mean_FC}{Only applies when thresh_method = "Mean"; The fold-change above the baseline for thresholding (2 by default).}

\item{proportion_max}{Only applies when thresh_method = "Max"; The proportion of Max fluorescence for thresholding (0.1 by default).}

\item{threshold}{Only applies when thresh_method = "Manual". Numerical value to be used as threshold.}

\item{thresh_calc_range}{Range in hours to be used as the basis for calculating threshold. By default thresh_calc_range = c(0,1).}
}
\value{
Outputs a matrix with the 1/lag values for each well with matching sample info data
}
\description{
calculates the inverse lag phase for each well of the QuIC plate (AKA amyloid formation rate)
}
\examples{
\dontrun{lag_data <- calc_lag_phase(signal_data, 40)}
\dontrun{lag_data <- calc_lag_phase(signal_data, 40, thresh_method = "2xMean", thresh_calc_range = c(13:16))}
\dontrun{lag_data <- calc_lag_phase(signal_data, 40, thresh_method = "Manual", threshold=15000)}
\dontrun{lag_data <- calc_lag_phase(signal_data, 40, thresh_method = "MaxSlope")}
}
