% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_functions.R
\name{calc_SD50}
\alias{calc_SD50}
\title{calculate SD50 values}
\usage{
calc_SD50(lag_data, starting_dilution = NULL, positivity_threshold = 0.75)
}
\arguments{
\item{lag_data}{Output of calc_lag_phase() function}

\item{starting_dilution}{The numeric value for the starting dilution used to calculate SD50. Useful if there is inhibition at higher sample concentrations.}

\item{positivity_threshold}{The proportion of positive wells in the first dilution must be above this threshold for a sample to be considered "positive". 0.75 by default}
}
\value{
Outputs SD50 values, standard error, and 95\% confidence interval for each sample with matching sample info. Samples deemed to be "negative" are returned with an SD50 value equivalent to 1 dilution lower than the minimum dilution.
}
\description{
Calculates SD50 for each sample from the lag phase data. 3/4 wells for the highest dilution must be
positive for the function to return SD50 measurements (This serves as the limit of detection).
Otherwise returns the maximum possible value below the limit of detection.
}
\examples{
sd50_res <- calc_SD50(lag_data)
sd50_res <- calc_SD50(lag_data, starting_dilution=1e-04)
}
